NJBuilder class, no trait, public associated function implementation
